+++ fortscale/fortscale-ml/src/main/java/fortscale/ml/model/builder/CategoryRarityModelWithFeatureOccurrencesDataBuilder.java
@@ -15,7 +15,7 @@ public class CategoryRarityModelWithFeatureOccurrencesDataBuilder extends Catego
 
     @Override
     public Model build(Object modelBuilderData) {
-        Map<String, Integer> featureValueToCountMap = castModelBuilderData(modelBuilderData);
+        Map<String, Long> featureValueToCountMap = castModelBuilderData(modelBuilderData);
         return new CategoryRarityModelWithFeatureOccurrencesData(getOccurrencesToNumOfFeatures(featureValueToCountMap));
     }
 }
